test: Move create_new_image() out to helpers.rs
